* move some stuff around * initialize_bytes failed when optimized and now it doesn't.
