+++ include/data/math/number/gmp/Z.hpp
@@ -5,10 +5,632 @@
 #ifndef DATA_MATH_NUMBER_GMP_Z
 #define DATA_MATH_NUMBER_GMP_Z
 
+#include <gmp.h>
+#include <gmpxx.h>
+
 #include <data/math/power.hpp>
 #include <data/math/number/gmp/mpz.hpp>
-#include <data/encoding/integer.hpp>
 #include <data/math/number/division.hpp>
+
+#include <data/math/number/bounded/bounded.hpp>
+#include <data/encoding/integer.hpp>
+#include <data/encoding/base58.hpp>
+
+namespace data::math::number::GMP {
+
+    typedef mp_limb_t gmp_uint;
+    typedef mp_limb_signed_t gmp_int;
+
+    // this is an impediment to working on Windows but we need it for now.
+    static_assert (sizeof (gmp_uint) == 8);
+    static_assert (sizeof (gmp_int) == 8);
+
+    const __mpz_struct MPZInvalid = __mpz_struct {0, 0, nullptr};
+
+    bool inline equal (const __mpz_struct &a, const __mpz_struct &b) {
+        return a._mp_alloc == b._mp_alloc && a._mp_size == b._mp_size && a._mp_d == b._mp_d;
+    }
+
+    uint32 inline size (const __mpz_struct &a) {
+        return a._mp_alloc;
+    }
+
+    bool inline valid (const __mpz_struct &mpz) {
+        return mpz._mp_d != nullptr;
+    }
+
+    math::sign inline sign (const __mpz_struct &mpz) {
+        return !valid (mpz) ? zero : math::sign {static_cast<int8_t> (mpz_cmp_si (&mpz, 0))};
+    }
+
+    template <endian::order r, size_t size, std::unsigned_integral word>
+    bool operator == (const Z &, const sint<r, size, word> &);
+
+    template <endian::order r, size_t size, std::unsigned_integral word>
+    std::strong_ordering operator <=> (const Z &, const sint<r, size, word> &);
+
+    template <endian::order r, size_t size, std::unsigned_integral word>
+    bool operator == (const N &, const uint<r, size, word> &);
+
+    template <endian::order r, size_t size, std::unsigned_integral word>
+    std::strong_ordering operator <=> (const N &, const uint<r, size, word> &);
+
+    template <endian::order r, neg c, std::unsigned_integral word>
+    bool operator == (const Z &, const Z_bytes<r, c, word> &);
+
+    template <endian::order r, neg c, std::unsigned_integral word>
+    std::strong_ordering operator <=> (const Z &, const Z_bytes<r, c, word> &);
+
+    template <endian::order r, std::unsigned_integral word>
+    bool operator == (const N &, const N_bytes<r, word> &);
+
+    template <endian::order r, std::unsigned_integral word>
+    std::strong_ordering operator <=> (const N &, const N_bytes<r, word> &);
+
+    struct Z final {
+        mpz_t MPZ;
+
+        Z ();
+        ~Z ();
+
+        Z (const Z &n);
+
+        Z (Z &&n);
+
+        Z &operator = (const Z &n);
+
+        Z &operator = (Z &&n);
+
+        // We need these to ensure that we can accept
+        // any number literal.
+        template <std::signed_integral I> Z (I);
+        template <std::unsigned_integral I> Z (I);
+
+        static Z read (string_view x);
+
+        Z (string_view);
+
+        Z (const dec_int &u): Z {string_view (u)} {}
+        Z (const dec_uint &u): Z {string_view (u)} {}
+        template <neg n, hex_case zz> Z (const hex::integer<n, zz> &);
+
+        template <endian::order r, neg c, std::unsigned_integral word>
+        explicit Z (const Z_bytes<r, c, word> &);
+
+        template <endian::order r, std::unsigned_integral word>
+        explicit Z (const N_bytes<r, word> &);
+
+        template <endian::order r, size_t size, std::unsigned_integral word>
+        explicit Z (const bounded<true, r, size, word> &x): Z {Z_bytes<r, neg::twos, word> {x}} {}
+
+        template <endian::order r, size_t size, std::unsigned_integral word>
+        explicit Z (const bounded<false, r, size, word> &x): Z {N_bytes<r, word> {x}} {}
+
+        size_t size () const;
+
+        mp_limb_t &operator [] (size_t i);
+
+        const mp_limb_t &operator [] (size_t i) const;
+
+        mp_limb_t *begin ();
+
+        mp_limb_t *end ();
+
+        const mp_limb_t *begin () const;
+
+        const mp_limb_t *end () const;
+
+        // TODO we can make a cast to any integral type
+        explicit operator int64 () const;
+        explicit operator int32 () const;
+        explicit operator uint64 () const;
+        explicit operator uint32 () const;
+        explicit operator double () const;
+
+        explicit operator dec_int () const;
+        template <hex_case zz> explicit operator hex::int2<zz> () const;
+        template <hex_case zz> explicit operator hex::intBC<zz> () const;
+
+        // TODO get rid of these operators and make them constructors.
+        template <endian::order r, neg c, std::unsigned_integral word>
+        explicit operator Z_bytes<r, c, word> () const;
+
+        template <endian::order r, size_t size, std::unsigned_integral word>
+        explicit operator bounded<true, r, size, word> () const {
+            return bounded<true, r, size, word> (this->operator Z_bytes<r, neg::twos, word> ());
+        }
+
+    };
+
+    // implementation of naturals given an implementation of integers.
+    // TODO N can extend Z.
+    struct N {
+        Z Value;
+
+        N () : Value {} {}
+
+        // need all of these to ensure that we can work with number
+        // literals.
+
+        // We need these to ensure that we can accept
+        // any number literal.
+        template <std::signed_integral I> N (I);
+        template <std::unsigned_integral I> N (I);
+
+        explicit N (const Z &z) : Value {z} {}
+        explicit N (Z &&z) : Value {z} {}
+
+        N (string_view);
+        N (const dec_uint &u): N {string_view (u)} {}
+        template <hex_case zz> N (const hex::uint<zz> &u): N {string_view (u)} {}
+        N (const base58_uint &u);
+
+        template <endian::order r, std::unsigned_integral word>
+        explicit N (const N_bytes<r, word> &n) : Value {n} {}
+
+        template <endian::order r, size_t size, std::unsigned_integral word>
+        explicit N (const bounded<false, r, size, word> &x): N {N_bytes<r, word> (x)} {}
+
+        operator Z () const {
+            return Value;
+        }
+
+        bool valid () const {
+            return data::valid (Value) && Value >= 0;
+        }
+
+        explicit operator double () const {
+            return double (Value);
+        }
+
+        explicit operator uint64 () const {
+            return uint64 (Value);
+        }
+
+        explicit operator uint32 () const {
+            return uint32 (Value);
+        }
+
+        explicit operator int64 () const {
+            return int64 (Value);
+        }
+
+        explicit operator int32 () const {
+            return int32 (Value);
+        }
+
+        explicit operator dec_uint () const;
+        explicit operator dec_int () const;
+        template <neg n, hex_case zz> explicit operator hex::integer<n, zz> () const;
+
+        template <endian::order r, std::unsigned_integral word>
+        explicit operator N_bytes<r, word> () const;
+
+        template <endian::order r, neg c, std::unsigned_integral word>
+        explicit operator Z_bytes<r, c, word> () const {
+            return Z_bytes<r, c, word> (N_bytes<r, word> (*this));
+        }
+
+        template <bool u, endian::order r, size_t size, std::unsigned_integral word>
+        explicit operator bounded<u, r, size, word> () const {
+            return bounded<u, r, size, word> (N_bytes<r, word> (*this));
+        }
+
+    };
+
+}
+namespace data::math::number::GMP {
+
+    inline Z::Z () {
+        mpz_init (MPZ);
+    }
+
+    inline Z::~Z () {
+        mpz_clear (MPZ);
+    }
+
+    template <std::signed_integral I> Z::Z (I x): Z {} {
+        mpz_init_set_si (MPZ, x);
+    }
+
+    template <std::unsigned_integral I> Z::Z (I x): Z {} {
+        mpz_init_set_ui (MPZ, x);
+    }
+
+    template <std::signed_integral I> N::N (I x): Value {} {
+        if (x < 0) throw exception {} << "N cannot be less than zero";
+        mpz_init_set_si (Value.MPZ, x);
+    }
+
+    template <std::unsigned_integral I> N::N (I x): Value {} {
+        mpz_init_set_ui (Value.MPZ, x);
+    }
+
+    inline Z::Z (const Z &n) : Z {} {
+        mpz_set (MPZ, n.MPZ);
+    }
+
+    inline Z::Z (Z &&n) : Z {} {
+        mpz_swap (MPZ, n.MPZ);
+    }
+
+    Z inline &Z::operator = (const Z &n) {
+        mpz_set (MPZ, n.MPZ);
+        return *this;
+    }
+
+    Z inline &Z::operator = (Z &&n) {
+        mpz_swap (MPZ, n.MPZ);
+        return *this;
+    }
+
+    size_t inline Z::size () const {
+        return GMP::size (MPZ[0]);
+    }
+
+    mp_limb_t inline &Z::operator [] (size_t i) {
+        if (static_cast<int> (i) >= MPZ[0]._mp_alloc) throw out_of_range {"Z"};
+        return *(MPZ[0]._mp_d + i);
+    }
+
+    const mp_limb_t inline &Z::operator [] (size_t i) const {
+        if (static_cast<int> (i) >= MPZ[0]._mp_alloc) throw out_of_range {"Z"};
+        return *(MPZ[0]._mp_d + i);
+    }
+
+    mp_limb_t inline *Z::begin () {
+        return mpz_limbs_modify (this->MPZ, mpz_size (this->MPZ));
+    }
+
+    mp_limb_t inline *Z::end () {
+        return mpz_limbs_modify (this->MPZ, mpz_size (this->MPZ)) + mpz_size (this->MPZ);
+    }
+
+    const mp_limb_t inline *Z::begin () const {
+        return mpz_limbs_read (this->MPZ);
+    }
+
+    const mp_limb_t inline *Z::end () const {
+        return mpz_limbs_read (this->MPZ) + mpz_size (this->MPZ);
+    };
+
+    inline Z::operator double () const {
+        return mpz_get_d (MPZ);
+    }
+
+    bool inline operator == (const Z &a, const Z &b) {
+        return a <=> b == 0;
+    }
+
+    bool inline operator == (const N &a, const N &b) {
+        return a <=> b == 0;
+    }
+
+    template <std::integral I> bool inline operator == (const N &a, I b) {
+        return a <=> b == 0;
+    }
+
+    template <std::integral I> bool inline operator == (I a, const N &b) {
+        return a <=> b == 0;
+    }
+
+    template <std::integral I> bool inline operator == (const Z & a, I b) {
+        return a <=> b == 0;
+    }
+
+    template <std::integral I> bool inline operator == (I a, const Z &b) {
+        return a <=> b == 0;
+    }
+
+    std::strong_ordering inline operator <=> (const N &a, const N &b) {
+        auto cmp = mpz_cmp (a.Value.MPZ, b.Value.MPZ);
+        return cmp < 0 ? std::strong_ordering::less :
+        cmp > 0 ? std::strong_ordering::greater : std::strong_ordering::equivalent;
+    }
+
+    std::strong_ordering inline operator <=> (const Z &a, const Z &b) {
+        auto cmp = mpz_cmp (a.MPZ, b.MPZ);
+        return cmp < 0 ? std::strong_ordering::less :
+        cmp > 0 ? std::strong_ordering::greater : std::strong_ordering::equivalent;
+    }
+
+    template <std::signed_integral I>
+    std::strong_ordering inline operator <=> (const Z &a, I b) {
+        auto cmp = mpz_cmp_si (a.MPZ, b);
+        return cmp < 0 ? std::strong_ordering::less :
+        cmp > 0 ? std::strong_ordering::greater : std::strong_ordering::equivalent;
+    }
+
+    template <std::signed_integral I>
+    std::strong_ordering inline operator <=> (I a, const Z &b) {
+        auto cmp = mpz_cmp_si (b.MPZ, a);
+        return cmp > 0 ? std::strong_ordering::less :
+        cmp < 0 ? std::strong_ordering::greater : std::strong_ordering::equivalent;
+    }
+
+    template <std::unsigned_integral I>
+    std::strong_ordering inline operator <=> (const Z &a, I b) {
+        auto cmp = mpz_cmp_ui (a.MPZ, b);
+        return cmp < 0 ? std::strong_ordering::less :
+        cmp > 0 ? std::strong_ordering::greater : std::strong_ordering::equivalent;
+    }
+
+    template <std::unsigned_integral I>
+    std::strong_ordering inline operator <=> (I a, const Z &b) {
+        auto cmp = mpz_cmp_ui (b.MPZ, a);
+        return cmp > 0 ? std::strong_ordering::less :
+        cmp < 0 ? std::strong_ordering::greater : std::strong_ordering::equivalent;
+    }
+
+    template <std::signed_integral I>
+    std::strong_ordering inline operator <=> (const N &a, I b) {
+        auto cmp = mpz_cmp_si (a.Value.MPZ, b);
+        return cmp < 0 ? std::strong_ordering::less :
+        cmp > 0 ? std::strong_ordering::greater : std::strong_ordering::equivalent;
+    }
+
+    template <std::signed_integral I>
+    std::strong_ordering inline operator <=> (I a, const N &b) {
+        auto cmp = mpz_cmp_si (b.Value.MPZ, a);
+        return cmp > 0 ? std::strong_ordering::less :
+        cmp < 0 ? std::strong_ordering::greater : std::strong_ordering::equivalent;
+    }
+
+    template <std::unsigned_integral I>
+    std::strong_ordering inline operator <=> (const N &a, I b) {
+        auto cmp = mpz_cmp_ui (a.Value.MPZ, b);
+        return cmp < 0 ? std::strong_ordering::less :
+        cmp > 0 ? std::strong_ordering::greater : std::strong_ordering::equivalent;
+    }
+
+    template <std::unsigned_integral I>
+    std::strong_ordering inline operator <=> (I a, const N &b) {
+        auto cmp = mpz_cmp_ui (b.Value.MPZ, a);
+        return cmp > 0 ? std::strong_ordering::less :
+        cmp < 0 ? std::strong_ordering::greater : std::strong_ordering::equivalent;
+    }
+
+    template <std::unsigned_integral I> N inline &operator /= (N &u, I x) {
+        return u &= N {x};
+    }
+
+    template <std::unsigned_integral I> N inline &operator %= (N &u, I x) {
+        return u &= N {x};
+    }
+
+    Z inline operator / (const Z &a, const Z &b) {
+        return def::divmod<Z, Z> {} (a, nonzero {b}).Quotient;
+    }
+
+    Z inline &operator /= (Z &a, const Z& z) {
+        return a = a / z;
+    }
+
+    Z inline operator ++ (Z &n, int) {
+        Z z = n;
+        ++ (n);
+        return z;
+    }
+
+    Z inline operator -- (Z &n, int) {
+        Z z = n;
+        ++ (n);
+        return z;
+    }
+
+    uint64 inline operator % (const Z &a, uint64 b) {
+        return uint64 (a % N (b));
+    }
+
+    template <endian::order r, size_t size, std::unsigned_integral word>
+    bool inline operator == (const Z &a, const sint<r, size, word> &b) {
+        return (a <=> b) == 0;
+    }
+
+    template <endian::order r, size_t size, std::unsigned_integral word>
+    bool inline operator == (const N &a, const uint<r, size, word> &b) {
+        return (a <=> b) == 0;
+    }
+
+    template <endian::order r, size_t size, std::unsigned_integral word>
+    std::strong_ordering inline operator <=> (const Z &a, const sint<r, size, word> &b) {
+        return a <=> Z (b);
+    }
+
+    template <endian::order r, size_t size, std::unsigned_integral word>
+    std::strong_ordering inline operator <=> (const N &a, const uint<r, size, word> &b) {
+        return a <=> N (b);
+    }
+
+    template <endian::order r, neg c, std::unsigned_integral word>
+    bool inline operator == (const Z &a, const Z_bytes<r, c, word> &b) {
+        return a == Z (b);
+    }
+
+    template <endian::order r, neg c, std::unsigned_integral word>
+    std::strong_ordering inline operator <=> (const Z &a, const Z_bytes<r, c, word> &b) {
+        return a <=> Z (b);
+    }
+
+    template <endian::order r, std::unsigned_integral word>
+    bool inline operator == (const N &a, const N_bytes<r, word> &b) {
+        return a == N (b);
+    }
+
+    template <endian::order r, std::unsigned_integral word>
+    std::strong_ordering inline operator <=> (const N &a, const N_bytes<r, word> &b) {
+        return a <=> N (b);
+    }
+
+    constexpr int inline endian_boost_to_GMP (endian::order r) {
+        return r == endian::order::big ? 1 : -1;
+    }
+
+    template <endian::order r, std::unsigned_integral word>
+    Z::Z (const N_bytes<r, word> &z) : Z {} {
+        mpz_import (
+            MPZ,
+            z.size (),
+                    endian_boost_to_GMP (r),
+                    sizeof (word),
+                    // TODO this may not be true in the future.
+                    endian_boost_to_GMP (endian::order::native),
+                    0, // number of unused bits in each element.
+                    z.data ());
+    }
+
+    template <endian::order r, neg c, std::unsigned_integral word>
+    Z::Z (const Z_bytes<r, c, word> &z) : Z {} {
+
+        if (data::is_zero (z)) return;
+
+        if (data::is_negative (z)) {
+            *this = std::move (Z (-z));
+            // this negates the value of the number.
+            MPZ[0]._mp_size = -MPZ[0]._mp_size;
+            return;
+        }
+
+        mpz_import (
+            MPZ,
+            z.size (),
+                    endian_boost_to_GMP (r),
+                    sizeof (word),
+                    // TODO this may not be true in the future.
+                    endian_boost_to_GMP (endian::order::native),
+                    0, // number of unused bits in each element.
+                    z.data ());
+    }
+
+    // TODO use mpz_export
+    template <endian::order r, neg c, std::unsigned_integral word>
+    Z::operator Z_bytes<r, c, word> () const {
+
+        if (data::is_negative (*this))
+            return -convert<Z_bytes<r, c, word>> (-(*this));
+
+        // the new number.
+        Z_bytes<r, c, word> nn {};
+        size_t n_size = mpz_size (MPZ);
+
+        if (n_size == 0) return nn;
+
+        if constexpr (sizeof (word) == sizeof (GMP::gmp_uint)) {
+            nn.resize (n_size + 1);
+            std::copy (begin (), end (), nn.words ().begin ());
+        } else if (sizeof (word) < sizeof (GMP::gmp_uint)) {
+            nn.resize (n_size * (sizeof (GMP::gmp_uint) / sizeof (word)) + 1);
+            auto nit = nn.words ().begin ();
+            for (const GMP::gmp_uint &limb : *this) {
+                GMP::gmp_uint z = limb;
+                for (int i = 0; i < sizeof (GMP::gmp_uint) / sizeof (word); i++) {
+                    *nit = static_cast<word> (z & std::numeric_limits<word>::max ());
+                    nit++;
+                    z >>= (sizeof (word) * 8);
+                }
+            }
+        } else if (sizeof (word) > sizeof (GMP::gmp_uint)) {
+            nn.resize (n_size / (sizeof (GMP::gmp_uint) / sizeof (word)) + 1);
+            auto nit = nn.words ().begin ();
+            auto b = begin ();
+            while (b != end ()) {
+                *nit = 0;
+                for (int i = 0; i < sizeof (GMP::gmp_uint) / sizeof (word); i++) {
+                    *nit += static_cast<word> (*b) << (i * sizeof (GMP::gmp_uint) * 8);
+                    b++;
+                }
+                nit++;
+            }
+        } else throw exception {} << "unhandled case in Z -> Z_bytes (this is impossible)";
+
+        // set last digit to zero.
+        *(nn.words ().end () - 1) = 0;
+
+        return trim (nn);
+    }
+
+    // TODO use mpz_export
+    template <endian::order r, std::unsigned_integral word>
+    N::operator N_bytes<r, word> () const {
+        // the new number.
+        N_bytes<r, word> nn {};
+        size_t n_size = mpz_size (Value.MPZ);
+        if (n_size == 0) return nn;
+
+        if constexpr (sizeof (word) == sizeof (GMP::gmp_uint)) {
+            nn.resize (n_size);
+            std::copy (Value.begin (), Value.end (), nn.words ().begin ());
+        } else if (sizeof (word) < sizeof (GMP::gmp_uint)) {
+            nn.resize (n_size * (sizeof (GMP::gmp_uint) / sizeof (word)));
+            auto x = nn.words ().begin ();
+            for (const GMP::gmp_uint &limb : Value) {
+                GMP::gmp_uint z = limb;
+                for (int i = 0; i < sizeof (GMP::gmp_uint) / sizeof (word); i++) {
+                    *x = static_cast<word> (z & std::numeric_limits<word>::max ());
+                    x++;
+                    z >>= (sizeof (word) * 8);
+                }
+            }
+        } else if (sizeof (word) > sizeof (GMP::gmp_uint)) {
+            nn.resize (n_size / (sizeof (GMP::gmp_uint) / sizeof (word)));
+            auto nit = nn.words ().begin ();
+            auto b = Value.begin ();
+            while (b != Value.end ()) {
+                *nit = 0;
+                for (int i = 0; i < sizeof (GMP::gmp_uint) / sizeof (word); i++) {
+                    *nit += static_cast<word> (*b) << (i * sizeof (GMP::gmp_uint) * 8);
+                    b++;
+                }
+                nit++;
+            }
+        } else throw exception {} << "unhandled case in N -> N_bytes (this is impossible)";
+
+        return trim (nn);
+    }
+
+    inline N::N (const base58_uint &u): Value {} {
+        if (!u.valid ()) throw exception {} << "invalid base 58 number" << *this;
+        *this = *encoding::base58::decode<N> (u);
+    }
+
+}
+
+namespace data::math::def {
+
+    Z inline identity<plus<Z>, Z>::operator () () {
+        return 0;
+    }
+
+    Z inline identity<times<Z>, Z>::operator () () {
+        return 1;
+    }
+
+    Z inline inverse<plus<Z>, Z>::operator () (const Z &a, const Z &b) {
+        return b - a;
+    }
+
+    Z inline times<Z>::operator () (const Z &a, const Z &b) {
+        return a * b;
+    }
+
+    nonzero<Z> inline times<Z>::operator () (const nonzero<Z> &a, const nonzero<Z> &b) {
+        return a * b;
+    }
+
+    N inline bit_xor<N>::operator () (const N &a, const N &b) {
+        return a ^ b;
+    }
+
+    Z inline bit_xor<Z>::operator () (const Z &a, const Z &b) {
+        return a ^ b;
+    }
+}
+
+namespace data::encoding::base58 {
+
+    template <std::integral I> inline string::string (I x): string {encode (N {x})} {}
+
+}
 
 namespace data::encoding::hexidecimal {
