+++ test/CMakeLists.txt
@@ -9,6 +9,11 @@ include (GoogleTest)
 
 add_executable (
     unit_tests
+    # we put these first even though they don't belong here because
+    # they take such a long time.
+    modular.cpp                  # TODO: A condition is commented out.
+    symmetric_crypto.cpp
+
     # These tests only require data_core for the components being tested.
     # other components may be used for the tests themselves.
     maybe_either.cpp
@@ -84,7 +89,7 @@ add_executable (
     sign.cpp                     # TODO: Contains commented test cases and should be a typed test
     abs.cpp                      # TODO: should be a typed test.
     roots.cpp                    # TODO: should be a typed test.
-    modular.cpp                  # TODO: A condition is commented out.
+    #modular.cpp                  # TODO: A condition is commented out.
     limits.cpp                   # TODO: cases are commented out.
     polynomial.cpp               # TODO: cases are commented out and should be a typed test.
     finite_field.cpp             # TODO: make typed test.
@@ -98,7 +103,7 @@ add_executable (
 
     # crypto
     # random_primes.cpp
-    symmetric_crypto.cpp
+    #symmetric_crypto.cpp
     NIST_DRBG.cpp
     secret_share.cpp
 )
